# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "媒体管理器 API 文档，包括媒体库管理、媒体文件管理等功能",
        "title": "媒体管理器 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:23347",
    "basePath": "/",
    "paths": {
        "/api/backup/cdn": {
            "post": {
                "description": "单独执行CDN文件的增量备份操作",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "备份管理"
                ],
                "summary": "手动触发CDN增量备份",
                "responses": {
                    "200": {
                        "description": "CDN备份任务已触发",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "触发失败或调度器未初始化",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/backup/database": {
            "post": {
                "description": "单独执行数据库备份操作，不包括CDN文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "备份管理"
                ],
                "summary": "手动触发数据库备份",
                "responses": {
                    "200": {
                        "description": "数据库备份任务已触发",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "触发失败或调度器未初始化",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/backup/history": {
            "get": {
                "description": "获取系统所有备份操作的历史记录，包括成功和失败的备份",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "备份管理"
                ],
                "summary": "获取备份历史记录",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量，默认20，最大100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "备份类型筛选：database | cdn | full",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controllers.BackupHistoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "获取失败或调度器未初始化",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/backup/restore/cdn/{backup_id}": {
            "post": {
                "description": "使用指定的备份记录恢复CDN文件到指定状态（谨慎操作）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "备份管理"
                ],
                "summary": "从备份恢复CDN文件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "备份记录ID（示例: 1）",
                        "name": "backup_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CDN文件恢复成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的备份 ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "备份记录不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "恢复失败或调度器未初始化",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/backup/status": {
            "get": {
                "description": "获取当前备份调度器和备份服务的运行状态信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "备份管理"
                ],
                "summary": "获取备份系统状态",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controllers.BackupStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "备份调度器未初始化",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/backup/trigger": {
            "post": {
                "description": "立即执行一次完整的系统备份，包括数据库和CDN文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "备份管理"
                ],
                "summary": "手动触发全量备份",
                "responses": {
                    "200": {
                        "description": "备份任务已触发",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "触发失败或调度器未初始化",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.BackupHistoryResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.BackupRecord"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 50
                }
            }
        },
        "controllers.BackupRecord": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "数据库定时备份"
                },
                "end_time": {
                    "type": "integer",
                    "example": 1640995260
                },
                "error": {
                    "type": "string",
                    "example": ""
                },
                "file_path": {
                    "type": "string",
                    "example": "/backups/db_20240101_120000.sql"
                },
                "file_size": {
                    "type": "integer",
                    "example": 1048576
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "start_time": {
                    "type": "integer",
                    "example": 1640995200
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "type": {
                    "type": "string",
                    "example": "database"
                }
            }
        },
        "controllers.BackupStatus": {
            "type": "object",
            "properties": {
                "failed_backups": {
                    "type": "integer",
                    "example": 2
                },
                "last_backup_time": {
                    "type": "integer",
                    "example": 1640991600
                },
                "next_backup_time": {
                    "type": "integer",
                    "example": 1640995200
                },
                "running": {
                    "type": "boolean",
                    "example": true
                },
                "successful_backups": {
                    "type": "integer",
                    "example": 23
                },
                "total_backups": {
                    "type": "integer",
                    "example": 25
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.ResponseCode"
                        }
                    ]
                },
                "data": {
                    "description": "数据"
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                },
                "timestamp": {
                    "description": "时间戳",
                    "type": "integer"
                }
            }
        },
        "response.ResponseCode": {
            "type": "integer",
            "enum": [
                200,
                400,
                401,
                403,
                404,
                405,
                409,
                422,
                429,
                500,
                501,
                502,
                503
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeBadRequest",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeMethodNotAllowed",
                "CodeConflict",
                "CodeUnprocessableEntity",
                "CodeTooManyRequests",
                "CodeInternalServerError",
                "CodeNotImplemented",
                "CodeBadGateway",
                "CodeServiceUnavailable"
            ]
        }
    },
...
}